CHROMIUM: security: Add container LSM proto definitions
Add proto definitions for generated events, container enlightenment and
configuration request/responses.

This change also contains the pre-generated source code and headers. It
removes the need for protoc support until upstream kernel potentially
adopts it. It explains why this code is not Linux style compliant and
further work is needed upstream to adapt the protoc plugin for it.

A quick README file that explains how to format proto files and update
nanopb generated files.

Protobuf enforces strict utf-8 checks on strings. If a string contains
non-utf8 characters, a message cannot be deserialized. To prevent that,
used bytes if the string is controlled by user-mode.

BUG=chromium:918980
TEST=Build, boot and GCP internal testing.

Signed-off-by: Thomas Garnier <thgarnie@google.com>
Signed-off-by: Matt Van Gundy <mattvangundy@google.com>
Change-Id: I3ff32b5c7fd29e38574ddd27dd14b58d0ebe2af6
Reviewed-on: https://chromium-review.googlesource.com/1379659
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>