UPSTREAM: ath10k: fixup extended per sta tx statistics
I had a couple concerns with this code.  First, we definitely need to
set the error code if the kzalloc() fails.  Secondly, I was worried that
if we didn't set "arsta->tx_stats" to NULL after freeing it, then it
looks to me like it might lead to a use after free.  I can't test that,
but it's harmless to set it to NULL so I did.

Fixes: a904417fc876 ("ath10k: add extended per sta tx statistics support")
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
(cherry picked from commit 539b44e70c5b4ad82d2c0c97a4dfb37523d891f6)

BUG=b:118512917
TEST=Run wifi_matfunc on the last commit.

Signed-off-by: Yen-lin Lai <yenlinlai@google.com>
Change-Id: I69b74a3a5790e8ed7ef41343d7986ce4a773d425
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2397923
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>