FROMLIST: Revert "drm/amd/powerplay: Enable/Disable NBPSTATE on On/OFF of UVD"
This reverts commit dbd8299c32f6f413f6cfe322fe0308f3cfc577e8.

Reason for revert:
This patch sends  msg PPSMC_MSG_DisableLowMemoryPstate(0x002e)
in wrong of sequence to SMU which is before PPSMC_MSG_UVDPowerON
(0x0008).
This leads to SMU failing to service the request as it is
dependant on UVD to be powered ON, since it accesses UVD
registers.

This msg should ideally be sent only when the UVD is about to decode
a 4k video.

Signed-off-by: Shirish S <shirish.s@amd.com>
(am from https://patchwork.freedesktop.org/patch/258514/)

BUG=b:117872984
TEST=On Careena dont see 0x002e msg sent before 0x0008

Change-Id: If4109c5ced19ef5971bf63acaa84ad4db928bcf7
Reviewed-on: https://chromium-review.googlesource.com/1299374
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Shirish S <shirish.s@amd.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>