FROMLIST: drm/amd/display: set backlight level limit to 1
This patch will work as workaround for silicon limitation
related to PWM dutycycle when the backlight level goes to 0.

Actually PWM value is 16 bit value and valid range from 1-65535.
when ever user requested to set this PWM value to 0 which is not
fall in the range, in VBIOS taken care this by limiting to 1.
This patch here will do the same. Either driver or VBIOS can not
pass 0 value as it is not a valid range for PWM and it will
give a high PWM pulse which is not the intended behaviour as
per HW constraints.

Signed-off-by: suresh guttula <suresh.guttula@amd.com>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>

(am from https://patchwork.freedesktop.org/patch/258862/)

BUG=b:113310483
TEST=Reduce the brightness level using F6 and check the panel
     of Careena EVT

Change-Id: Iba97f902eb45b73e3b8c9799cf37b197d45d36a3
Reviewed-on: https://chromium-review.googlesource.com/1301676
Commit-Ready: Akshu Agrawal <akshu.agrawal@amd.com>
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Akshu Agrawal <akshu.agrawal@amd.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>