CHROMIUM: iwl7000: mac80211: Properly handle SKB with radiotap only
The monitor interface Rx handling of SKBs that contain only
radiotap information was buggy as it tried to access the
SKB assuming it contains a frame.

To fix this, check the RX_FLAG_NO_PSDU flag in the Rx status
(indicting that the SKB contains only radiotap information),
and do not perform data path specific processing when the flag
is set.

Change-Id: Ib9bb43f584b7d6fed8602575913333744e07c9eb
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: c195cc5118b561f94eab03e402a456ca59d40870