Revert "CHROMIUM: ASoC: DA7219: For Grunt save and restore state of register on S3"
This reverts commit c27e1d170103eaebed487027925038038c755dd5.

The hw fix is in for all platform and the CL was for only Grunt EVT.
But we see this piece of code is executed even for other platforms.

In S3 cycles a fail i2c transaction in resume can lead to storing of
incorrect value and thus lead to restore of incorrect state.

BUG=b:118248953, b:111745987
TEST= audio works after S3

Change-Id: I85dee8f71bbd1eb9ede6a961ed801defbd53c1ff
Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>
Reviewed-on: https://chromium-review.googlesource.com/1360296
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>