FROMLIST/BACKPORT: drm/amd/display: Warn instead of error on REG_WAIT timeout
[Why]
DC warns when a REG_WAIT takes a while and full-on errors
with stack dump on REG_WAIT timeout.  Most of the time it isn't
a real issue.

[How]
Make DC cool its jets - taking a while is a debug message (because
it is not something that normal users should need to be aware of),
and timeouts are warnings (because it technically shouldn't
happen, but it's not a big deal if it does)

Signed-off-by: David Francis <David.Francis@amd.com>
Reviewed-by: Harry Wentland <Harry.Wentland@amd.com>
Acked-by: Leo Li <sunpeng.li@amd.com>

(am from https://patchwork.freedesktop.org/patch/267613/)

BUG=b:120407715
TEST=Build & boot.
Signed-off-by: Shirish S <shirish.s@amd.com>

Conflict in drivers/gpu/drm/amd/display/dc/dc_helper.c:
	reuse dm_logger_write instead of backporting DC_LOGGER_WRITE

Change-Id: I1a4e2785a3aded2d847ce4c3da260d4d2a3edd2a
Reviewed-on: https://chromium-review.googlesource.com/1379833
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>