CHROMIUM: Input: atmel_mxt_ts - Move driver from 4.4.166
Takes the atmel_mtx_ts driver from commit 47e0272c25ee ("CHROMIUM: Merge
'v4.4.157' into chromeos-4.4") and clobbers the existing one. This is
needed to get the touch screen working on samus after suspend/resume.

This has the effect of nuking the history for this particular file since
it diverged in commit 151a8d0d532d ("CHROMIUM: For 3.18 rebase of Atmel
changes, start with 3.14."), but this is favorable to a rebase.

BUG=b:120088009
TEST=close/open lid and verify that the touch screen responds

Change-Id: I4cbd760547f6e23a3a4b556cd970c6a2a3d6098b
Reviewed-on: https://chromium-review.googlesource.com/1378959
Commit-Ready: Jon Flatley <jflat@chromium.org>
Tested-by: Jon Flatley <jflat@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Ross Zwisler <zwisler@chromium.org>