CHROMIUM: iio: cros_ec: do not use flush packets for batch spreading
Flush events happen asynchronously to the regular flow of sensor
events, and as such they do not make good candidates to be used as
the basepoint to calculate sample periods for spreading.

This patch prevents flush samples to be inserted in a batch for spreading
as well as being used as the start of a batch.

BUG=b:120508077
TEST=observe flush events pushed to IIO but not spread

Change-Id: I78e964e2214ea3fcaa5fe3b177f6567ed94ce7e5
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1380973
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
(cherry picked from commit c59507474254b27ee9d8100e7c1e1be653224dd9)
Reviewed-on: https://chromium-review.googlesource.com/1415433