CHROMIUM: config: base: make CONFIG_HID_MULTITOUCH builtin
On Octopus devices, it appears that the hid-multitouch module must be
loaded for the synaptics device to appear properly under /dev/hidraw*.
Without this change, the synaptics updater will run before the
hid-multitouch module is loaded, and it will fail to update.

Other kernel versions enable this CONFIG option to improve bluetooth
behavior.  Here's one example:
  https://chromium-review.googlesource.com/599272

BUG=b:110761339
TEST=None

Signed-off-by: Justin TerAvest <teravest@chromium.org>

Change-Id: I95ac7307bb1d9a0a9ed57fc66f8b097c42adc044
Reviewed-on: https://chromium-review.googlesource.com/c/1387625
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>