CHROMIUM: ASoC: SOF: reduce timeout for ROM init and modify FW load retries
There is a possibility of CSE auth failing during stress tests
due to a race condition between when the audio driver requests
fw authentication and then the CSE is ready. Reduce the timeout
to check for ROM init and retry this step a few times to
improve the chances of FW download being successful.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>

BUG=b:110025789
TEST=suspend stress test passes and doesn't fail due to fw load error

Change-Id: I486e1361dea066405e2a14f6c0e58b2e45ebbad8
Signed-off-by: Naveen Manohar <naveen.m@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/1392727
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Naveen M <naveen.m@intel.corp-partner.google.com>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>