FROMGIT: ASoC: da7219: MCLK should be enabled before DAI clocks
For platforms using the Common Clock Framework to control the
codec's DAI clocks, MCLK should be enabled prior to DAI clocks
being turned on. For some platforms the codec is already
provided with an MCLK reference and can therefore control MCLK
itself as it needs to.

To improve functionality MCLK is now added as a parent to the
DAI clocks, if MCLK was provided, so that if they are enabled MCLK
will automatically be enabled as a prerequisite by the CCF.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit a6028cc60aad18d5d7c25d99b5cb8c24399387c3
https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git)
Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>

BUG=b:119926436
TEST=cras_test_client --capture_f <file> --duration_seconds 1
	cras_test_client --playback_f <file>
	=> No "pop" or "bo" sound at start of capture.

Change-Id: Ica84f4f91df5928353ecc78c2db7e2543bab96e3
Reviewed-on: https://chromium-review.googlesource.com/1404655
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Akshu Agrawal <akshu.agrawal@amd.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>