CHROMIUM: drm/i915: Set enable_psr param as safe
The i915.enable_psr parameter was marked as "unsafe" in
commit 25e1793f6c64 ("drm/i915: Mark debug mod options as _unsafe").
This was done upstream because of a number of bug reports that
suggested this feature can cause poor performance and system hangs
for reasons that were never diagnosed.

Broadwell uses this feature seemingly without issue so the "unsafe"
declaration can be dropped to avoid kernel tainting.

BUG=b:121042222
TEST=build + deploy kernel and verify no kernel tainting in dmesg

Change-Id: I41fcbc2e1a1ea3ecd297b535d97276374c762281
Signed-off-by: Jon Flatley <jflat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1405591
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Ross Zwisler <zwisler@chromium.org>