FROMLIST: ASoC: soc-core: fix init platform memory handling
snd_soc_init_platform initializes pointers to snd_soc_dai_link which is
statically allocated and it does this by devm_kzalloc. In the event of
an EPROBE_DEFER the memory will be freed and the pointers are left
dangling. snd_soc_init_platform sees the dangling pointers and assumes
they are pointing to initialized memory and does not reallocate them on
the second probe attempt which results in a use after free bug since
devm has freed the memory from the first probe attempt.

Since the intention for snd_soc_dai_link->platform is that it can be set
statically by the machine driver we need to respect the pointer in the
event we did not set it but still catch dangling pointers. The solution
is to add a flag to track whether the pointer was dynamically allocated
or not.

Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
(am from http://mailman.alsa-project.org/pipermail/alsa-devel/2019-January/144148.html)

BUG=b:110025789
TEST=test basic audio functionality on samus, octopus and grunt

Conflicts:
	sound/soc/soc-core.c
	Depends on UPSTREAM: ASoC: soc-core: manage platform name under snd_soc_init_platform()
	Which depends on UPSTREAM: ASoC: Remove platform code now everything is componentised
	Which is what this patchset is preparing to land. Will revert
	and reland this cleanly once the platform code is removed

Change-Id: I2ce2291566aac77804efaedcdb21beacb58147ad
Reviewed-on: https://chromium-review.googlesource.com/1405813
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>