CHROMIUM: iwl7000: mvm: fix civic request in FTM range request
The flag that indicates to request CIVIC info in the range request
was not set correctly. As a result, CIVIC was never reqeusted.
Fix it.

In addition, the tsf_mac_id in the request is not set, thus contains
some garbage value that may confuse the fw. Set it to 0xff to indicate
TSF reporting is not needed.

Change-Id: I3ce14c14410c64615321a864b5c2e8ecd7ff8abd
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
iwl7000-tree: 41029768b04a2c0928eb0af76ef7ebc09cd75e32