CHROMIUM: iwl7000: mvm: fix the spatial reuse parsing for HE_TRIG PPDUs
The spatial reuse 4 words fields are fetched from the HE-SIGA
by the firmware and propagated to the driver through the
Rx info. This is useful to populate the radiotap header.

We were looking at the wrong place in the firmware data and
got bogus values. Fix that.

Change-Id: I4bddb1e2069ee233968a81668b4e2a8c333e28c8
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: a613bf3b758470c8bf5f01536f9b4f7de4878129