CHROMIUM: iwl7000: nvm-parse: advertise IEEE80211_VHT_EXT_NSS_BW_CAPABLE in VHT
We support this, so we should advertise it. In fact, if we don't,
mac80211 will do it for us (as we advertise SUPPORTS_VHT_EXT_NSS_BW
to it), but that requires a memory reallocation which is wasteful:

  ieee80211 phy0: copying sband (band 1) due to VHT EXT NSS BW flag

Set the flag here to avoid that.

Change-Id: If9612b277a1bbe0d75ab1fe8070817bb7d14740e
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: da7e2ee819ded6d0ac0e0bf3896267edb66756b7