Revert "x86,kvm: move qemu/guest FPU switching out to vcpu_run"
With this commit included, crosvm on v4.14 host kernels experiences
non-deterministic guest memory corruption.  One example of the way this
can reproduce is with a message like this while starting init:

  Inconsistency detected by ld.so: dl-version.c: 205:
  _dl_check_map_versions: Assertion `needed != NULL' failed!

The underlying issue isn't understood yet, but for now, let's revert the
commit that exposes the problem and investigate further.

This reverts commit 4124a4cff344abbf8187775eb643d9827830e715.

BUG=chromium:925063
TEST=start termina on grunt (v4.14 host kernel)

Change-Id: If7c6223edf93eaa467f922306a5f26edf012cbe2
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1436220
Reviewed-by: Stephen Barber <smbarber@chromium.org>