CHROMIUM: cros_ec: use ODR events to clear the spreading history
This uses the ODR flag introduced in CL:1450288 to clear
the last batch data used by the spreading algorithm.

The spreading code relies on accurately measuring the effective
interval between batches of samples. An ODR change means that the
previous batches may no longer represent the behavior of the EC.
If an ODR change is immediately followed by a batch, the AP
will attempt to spread based on the previous interval, which
can easily lead to "spreading into the future", leading to
subsequent timestamp mismatch errors.

This patch solves that problem by having ODR events be breaking
points for the spreading code. The tradeoff is that this can
cause samples to be lost immediately after an ODR change.

This problem can't effectively be solved without either:
- storing those samples in a side queue and spreading them once
  updated information becomes available to do so;
- communicating the ODR to the AP as part of the ODR event.

BUG=b:123700100
TEST=running with CL:1450791, observe the
ring not using stale batch information
CQ-DEPEND=CL:1450288

Change-Id: Ic73d1bf0e7f659210052c56fe8d265f82f7f4b58
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1450743
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>