HACK: CHROMIUM: cros_ec: Support MKBP sync IRQ defined via ACPI GpioInt{}
On some ChromeOS devices, the EC -> AP sync IRQ is registered using a
GpioInt{} ACPI command. This is done in order to allow routing via
the AMD GPIO controller.

Retrieving the IRQ number in this case requires a different code path that
goes via ACPI and a GPIO-to-IRQ conversion instead of using the platform
device to directly retrieve the IRQ number.

We are working with upstream to get
https://patchwork.kernel.org/patch/10801897/ merged, but until that
reaches consensus, this local hack enables us to make forward progress.

BUG=b:123750725
TEST=observe chromeos-ec interrupt in /proc/interrupts

Change-Id: Ic88d28c4c921a5520759640a9fb0c203ec05e2c8
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1455372
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>