CHROMIUM: iwl7000: mvm: don't rely on mvm in the IWL_MVM_D3_DEBUG macro
In the upstream code, we don't have the debug override macros, so we
get a defined but not used warning.  To solve that, remove the local
declaration of mvm and put it inside the macro (with the proper cast)
so it doesn't show up in the upstream code.

Change-Id: Iff2a4218eb4d8dbc132f6eaabe7b7e63e6a84c5b
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: c95539be824e999432eb03d615a3841323d98172