FROMLIST: [v2] driver: platform: Support parsing GpioInt 0 in platform_get_irq()
ACPI 5 added support for GpioInt resources as a way to provide
information about interrupts mediated via a GPIO controller.

Several device buses (e.g. SPI, I2C) have support for retrieving
an IRQ specified via this type of resource, and providing it
directly to the driver as an IRQ number.

This is not currently done for the platform drivers, as platform_get_irq()
does not try to parse GpioInt() resources. This requires drivers to
either have to support only one possible IRQ resource, or to have code
in place to try both as a failsafe.

While there is a possibility of ambiguity for devices that exposes
multiple IRQs, it is easy and feasible to support the common case
of devices that only expose one IRQ which would be of either type
depending on the underlying system's architecture.

This commit adds support for parsing a GpioInt resource in order
to fulfill a request for the index 0 IRQ for a platform device.

from: https://patchwork.kernel.org/patch/10806705/
BUG=b:123750725
TEST=boot Kasumi360, observe GPIO work
Signed-off-by: Enrico Granata <egranata@chromium.org>

Change-Id: I9d106207932588ad587c32201f63841c92b7f7fd
Reviewed-on: https://chromium-review.googlesource.com/1459156
Commit-Ready: Enrico Granata <egranata@chromium.org>
Tested-by: Enrico Granata <egranata@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>