CHROMIUM: intel_ddi: Don't check polarity on eDP
With the weirdness of USB Type-C on samus we need to check for port
polarity in the i915 driver. Currently this polarity check happens
prior to lane training on all ports, including eDP ports. This causes
the cros_ec_pd_update code that is called to check polarity to throw an
error.

To solve this simply check if the port is eDP before querying polarity.

TEST=check that dmesg does not throw drm errors during boot + verify
     that external DP works on all ports in all orientations on samus
BUG=b:124880987

Change-Id: I7de11c8af4ced956e51a31956d9a640b587a89ed
Signed-off-by: Jon Flatley <jflat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1479552
Reviewed-by: Ross Zwisler <zwisler@chromium.org>