CHROMIUM: iwl7000: chromeOS: implement set/get iftypes functions for >= 5.0
We were not defining the ieee80211_sband_set_num_iftypes_data()
function and friends if the kernel was >= 4.21.  But we actually need
to define them so they set and get the data in the new structures.
Fix that and while at it, use 5.0 instead of 4.21.

Change-Id: I829b19062a68f81d496218ebcbff67896eadc2a0
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 1143f9fb0cc97f16e54d5537772169c1bd12b1ba