Revert "UPSTREAM: Bluetooth: btusb: driver to enable the usb-wakeup feature"
This reverts commit 74f911a25030 ("UPSTREAM: Bluetooth: btusb: driver to
enable the usb-wakeup feature"). The reason is that this causes any inband
communication from the btusb device to cause the system to wake up. We
know that the BT controller could be scanning when we decide to suspend
and will send in the results once the scan is complete, thus failing the
suspend entry.

Bluetoothd should be listening to powerd for suspend requests to handle
this properly, but we've observed enough spurious wakes that something
still isn't right.

(Reverting this commit has the side effect that BLE mouse clicks do not
wake up the system from suspend, but that is lesser price to pay than no
going to suspend).

This was previously done for 4.4 at:
  https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/750073

Signed-off-by: Rajat Jain <rajatja@google.com>
Signed-off-by: Justin TerAvest <teravest@chromium.org>
BUG=b:128319712
TEST=system can go to suspend even though BLE mouse is connected / scan
is running

Change-Id: Ia40208bf3ba45d722e754c49dead41a8ccf4a519
Reviewed-on: https://chromium-review.googlesource.com/1520811
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Yoni Shavit <yshavit@chromium.org>