FROMGIT: ASoC: da7219: Expose BCLK and WCLK control through CCF
For the purposes of platforms which use the codec as DAI clock
master for the CPU and other codec devices, there is the need to
not only expose the clock gating of BCLK and WCLK but also the
ability to set those rates without going through the ASoC APIs.

To make this possible, the previous CCF implementation in the
driver has been extended to separate BCLK and WCLK out. WCLK is
the parent clock to BCLK, and is also the clock gate for both.
BCLK in HW is a factor/multiplier of WCLK so derives from whatever
SR is chosen for WCLK, hence the need to make it a child of WCLK
for the purposes of CCF. Enabling/disabling either BCLK or WCLK
will result in clocks being ungated/gated accordingly. To simplify
matters, these clocks can only be configured if the codec is set
as master, otherwise CCF control is disallowed.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit d90ba6c8b53e541913a181638c353cf7a0856256
https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git for-next)
Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>

BUG=b:123433680
TEST=Ran audio and check for 64 bclks per wclk
cat /sys/kernel/debug/regmap/i2c-DLGS7219\:00/registers | grep 2b
2b: 81

Change-Id: Icfad6a9d5c78400135aedc1c4cb5c1d79679b63d
Reviewed-on: https://chromium-review.googlesource.com/1571058
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Akshu Agrawal <akshu.agrawal@amd.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>