CHROMIUM: security: Add more kernel command-line options.
To allow more flexibility in the features used by the LSM, add kernel
command-line arguments to enable or disable vsock and the sysfs
configuration which was previously using a CONFIG option.

The sysfs configuration option was previously only used for debug but
as we will have environments without vsock support, we might enable it
at runtime.

BUG=b/130802182
TEST=Tested enabled/disabled state and ran all existing tests.

Signed-off-by: Thomas Garnier <thgarnie@chromium.org>
Change-Id: I45d56a3bfd73064a84e66aa861a0fcc627307cab
Reviewed-on: https://chromium-review.googlesource.com/1576062
Reviewed-by: Guenter Roeck <groeck@chromium.org>