UPSTREAM: ASoC: rt5677-spi: Handle over reading when flipping bytes
There is a case when a we want to read a large number of bytes that
require a burst but is not a multiple of the word size (8). When this
happens rt5677_spi_reverse will run off the end of the buffer. The
solution is to tell spi_reverse the actual size of the destination and
stop if we reach it even if we have data left that we read.

Cc: Ben Zhang <benzh@chromium.org>
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 7b8164c1a29ce8ef91672c50ceac5c14475f5601)

TEST=test jack detect and hotwording after series is applied
BUG=b:139552675

Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Change-Id: Ib5e23e074a1b209859e3edbc1799756331f6c764
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1918532