CHROMIUM: Revert "mm: remove optimizations based on i_size in mapping writeback waits"
This reverts commit ffb959bbdf923b4f89a08a04aba2501b1b16d164.

commit ffb959bbdf92 ("mm: remove optimizations based on i_size in
  mapping writeback waits")

was itself a revert of a much older commit which optimized calls to
filemap_fdatawait_keep_errors() so that the range being waited upon was
limited to the current size of the inode.  This original commit was
reverted because of concerns that the optimization may be racy,
especially with respect to clustered or network filesystems.

However, we are currently seeing an issue where we wait excessively in
filemap_fdtawait_keep_errors().  Lots of details can be found here:

https://www.spinics.net/lists/linux-block/msg38204.html

This excessive wait ultimately results in the system rebooting due to
a hung task timeout.

Using the optimization where we wait based on the size of the inode
resolves this hung task issue for Chrome OS.  So, use that unless/until
we have a better fix based on the scope of jbd2 journal transactions or
something.  This should be relatively low risk as clustered/networked
filesystems are a less common use case for Chrome OS.

BUG=b:130521177, chromium:959968
TEST=DD'd 16 GiBs from /dev/zero to /mnt/stateful_partition, and was
able to copy a 15 GiB file from a USB stick to Downloads via the Files
app.  This was all done on a Bobba.

Change-Id: If3342d2bd02e68f511b6d1a1a14bc696b07fbfbc
Signed-off-by: Ross Zwisler <zwisler@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1632121
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Justin TerAvest <teravest@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Yu Zhao <yuzhao@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>