Revert "CHROMIUM: Revert "mm: remove optimizations based on i_size in mapping writeback waits""
This reverts commit b96b564105d53e4318b34715577a6ba44aef901b.

The above mentioned commit was a temporary fix which optimized calls to
filemap_fdatawait_keep_errors() so that the range being waited upon was
limited to the current size of the inode.  This prevented excessive
waits as described here:

https://www.spinics.net/lists/linux-block/msg38204.html

and in the bugs in the BUG line below.  However, we now have a real fix
which is destined for upstream:

https://patchwork.kernel.org/cover/11007125/

so use that instead.

BUG=b:130521177, chromium:959968, b:133470926, b:65528573
TEST=Multiple iteration of dd from /dev/zero to /mnt/stateful_partition,
filling the disk.  Also, copied multi-GiB files from USB sticks to the
Downloads folder via the GUI.  I did this on bobba, sarien and
samus-kernelnext.

Cq-Depend: chromium:1669281
Change-Id: Ic020d7f681bdb5f2b862a40720b088d0daf10f15
Signed-off-by: Ross Zwisler <zwisler@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1669278
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>