UPSTREAM: ASoC: rt5677: clear interrupts by polarity flip
The rt5677 jack detection function has a requirement that the polarity
of an interrupt be flipped after it fires in order to clear the
interrupt.

This patch implements an irq_chip with irq_domain directly instead of
using regmap-irq, so that interrupt source line polarities can be
flipped in the irq handler.

The reason that this patch does not add this feature within regmap-irq
is that future patches will add hotword detection support to this irq
handler. Those patches will require adding additional logic that would
not make sense to have in regmap-irq.

Signed-off-by: Ben Zhang <benzh@chromium.org>
Signed-off-by: Fletcher Woodruff <fletcherw@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 4f7b018b55db0361718161e1471d8b7a16fdfa47)

TEST=test jack detect and hotwording after series is applied
BUG=b:139552675

Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Change-Id: I67ab7a5b2f2577cd277dcbd4a5afd70ec34174bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1918525
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>