UPSTREAM: ASoC: rt5677: handle concurrent interrupts
The rt5677 driver writes to the IRQ control register within the IRQ
handler in order to flip the polarity of the interrupts that have been
signalled.  If an interrupt fires in the interval between the
regmap_read and the regmap_write, it will not trigger a new call to
rt5677_irq.

Add a bounded loop to rt5677_irq that keeps checking interrupts until
none are seen, so that any interrupts that are signalled in that
interval are correctly handled.

Signed-off-by: Ben Zhang <benzh@chromium.org>
Signed-off-by: Fletcher Woodruff <fletcherw@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit df9091e9d3f4500bc6fb15f5d2a1c2614f67004c)

TEST=test jack detect and hotwording after series is applied
BUG=b:139552675

Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Change-Id: I248aaeefa4680de37f99d4a696486117ea6d6093
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1918526
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>