CHROMIUM: iwl7000: mvm: enable Atlas platform workaround
Optionally detect if running on an Atlas platform, and if so tell
the firmware to work around the hardware issues there.

Due to the urgent nature of this request, we didn't yet design a
better solution, this will come in the future and then will not
just use a currently ignored bit.

Change-Id: I7b44eb9f65d4154bc5d479e4f438635c096135af
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: baa8c4d32090d0f29e78ccb574d8f590499d5516