CHROMIUM: iwl7000: tree-wide: set genl non-strict validation flags
Due to the decision to forward port our driver in the sense that
our driver is now based on an older kernel version and hasn't
merged upstream back yet, but is declared to be compatible with
running on 5.2 despite this, we run into the problem that all the
generic netlink ops are not setting any of the flags for non-
strict validation.

If running on 5.2, the kernel will validate all attributes for any
new commands (that aren't setting flags otherwise) *strictly*, and
this fails with iw not setting NLA_F_NESTED on nested attributes.

The "proper" fix for this would have been to not ship an old driver
to a newer kernel than that it's based on, but that didn't happen
due to time constraints etc.

To fix this now, we need to add the required non-strict validation
flags. This happened upstream as well, but since we don't have a
proper backport for generic netlink yet, that wouldn't compile, so
for now we need to put them under #ifdef.

This was done with a similar spatch (except for the ifdef) as the
one I used upstream to make the changes to start with:

    @@
    identifier ops;
    expression X;
    @@
    struct genl_ops ops[] = {
    ...,
     {
            .cmd = X,
    +#if LINUX_VERSION_IS_GEQ(5,2,0)
    +       .validate = GENL_DONT_VALIDATE_STRICT | GENL_DONT_VALIDATE_DUMP,
    +#endif
            ...
     },
    ...
    };

Change-Id: Ic563a4d55a1b18c11abc348c11e798697d42bdaa
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 905157a42471efef092912bb752c2c0f3bd3c25c