CHROMIUM: iwl7000: dbg_ini: load external dbg cfg after internal cfg is loaded
In the new API implementation the driver does not keep the internal and
external debug configurations in separate structs so the last
configuration that is loaded overrides the previous ones (this is true
only in some of the TLVs e.g. the buffer allocation TLV).
Load the external configuration after the internal is loaded so that
user configuration will override the default coming with the FW.

Change-Id: Ieb918aa866eef6e610956fb47c3b4dd31278fb7f
Signed-off-by: Shahar S Matityahu <shahar.s.matityahu@intel.com>
iwl7000-tree: f874878162180e8e08ed27de54fc04b574fa60c4