CHROMIUM: iwl7000: mvm: fix an out-of-bound access
The index for the elements of the ACPI object we dereference
was static. This means that if we called the function twice
we wouldn't start from 3 again, but rather from the latest
index we reached in the previous call.
This was dutifully reported by KASAN.

Fix this.

Change-Id: If76baed86a46464f3448d5957978ec446e5d3559
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 29e4edc65e6ba21f22162cc08e95f7b0066f93c0