UPSTREAM: lib/timerqueue: Rely on rbtree semantics for next timer
Simplify the timerqueue code by using cached rbtrees and rely on the tree
leftmost node semantics to get the timer with earliest expiration time.
This is a drop in conversion, and therefore semantics remain untouched.

The runtime overhead of cached rbtrees is be pretty much the same as the
current head->next method, noting that when removing the leftmost node,
a common operation for the timerqueue, the rb_next(leftmost) is O(1) as
well, so the next timer will either be the right node or its parent.
Therefore no extra pointer chasing. Finally, the size of the struct
timerqueue_head remains the same.

Passes several hours of rcutorture.

BUG=b:202231721
TEST=CQ

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lkml.kernel.org/r/20190724152323.bojciei3muvfxalm@linux-r8p5
(cherry picked from commit 511885d7061eda3eb1faf3f57dcc936ff75863f1)
Signed-off-by: Zubin Mithra <zsm@chromium.org>
Change-Id: Icad0d355bcf3a9abb9e2e932d62707d0287f0a00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3208669
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Martin Faltesek <mfaltesek@google.com>