CHROMIUM: iwl7000: add umac_prph_offset to trans-specific configuration
When the OTP is blank, we need to read from the UMAC prph registers in
order to find out the RF ID and decide which firmware (and other
configurations) to use.  The umac_prph_offset element was not in the
trans-specific configuration and we would try to access it when the
generic cfg was not set yet, causing a null-pointer dereference.

Fix this by moving umac_prph_offset to the trans-specific
configuration structure.

Change-Id: Ic412dfac4831aec9ff862cce6347fe7fe590f203
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 28ae90b6bbffbde9ddf94acba2035fffaf5c54c0