FROMLIST: rtw88: pci: reset H2C queue indexes in a single write
If the driver doesn't reset the host's and device's indexes
in a single write, the indexes will become different in a
short period. And it will confuse the DMA engine, make it
start to process non-existed entries.

Better to Write-1-to-reset the indexes, for the DMA engine
to know that this is a reset of the H2C queue, not a kick
off.

Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
(am from https://patchwork.kernel.org/patch/11146419/)
(also found at https://lkml.kernel.org/r/1568617425-28062-4-git-send-email-yhchuang@realtek.com)

BUG=b:124400293
TEST=see if there's any DMA error

Change-Id: Ib2f4904f0533351737b534f54375e228e2356b10
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Reviewed-on: https://chromium-review.googlesource.com/1775693
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Ready: Brian Norris <briannorris@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>