CHROMIUM: iwl7000: fw: remove unnecessary and out-of-place log
The log says that we're sending the command in a function that doesn't
send the command.  Additionally, logging that we're sending commands
is unnecessary, because we can both see it with trace-cmd and with
hcmd logs if they are enabled.

Change-Id: I4a53bf49a140f8733a8485756105e2fc66b032e3
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: c9f375adbd4c59685627f48cccf0b91e251e6161