CHROMIUM: iwl7000: mac80211: simplify TX aggregation start
There really is no need to make drivers call the
ieee80211_start_tx_ba_cb_irqsafe() function and then
schedule the worker if all we want is to set a bit.

Add a new return value (that was previously considered
invalid) to indicate that the driver is immediately
ready for the session, and make drivers use it. The
only drivers that remain different are the Intel ones
as they need to negotiate more with the firmware.

Change-Id: I152912660131cbab2e5d80b4218238c20f8a06e5
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: b0cb167d573cd0031ddee20a5d43afef49aec6da