CHROMIUM: usb: host: fix USB 3.0 breakage
As reported upstream via this conversation:

https://lore.kernel.org/linux-usb/CAGRrVHzta7YfzZYuc3d_5CS_SyG+at_=-T7Sm-DG-AS_ejsWug@mail.gmail.com/

the XHCI_SPURIOUS_WAKEUP quirk is destructive for BDW based Chromeboxes.
If the system is powered off and then immediately powered on, this quirk
will cause the USB 3.0 hub to be unable to communicate with all USB 3.0
devices for the duration of the next boot.

This quirk has the same behavior in the older v3.14 kernel, but wasn't
enabled.  It was turned on via this commit:

commit fd7cd061adcf ("xhci: Add spurious wakeup quirk for LynxPoint-LP controllers")

which was first merged in v4.3.  Unblock guado-kernelnext for now by
turning off the quirk.  Longer term we should implement a better fix,
such as having the USB startup code be able to deal with the fact that
the USB hub might be in D3-hot state.

BUG=b:140826284
TEST='halt -p' on guado-kernelnext, restart system immediately.  Verify
that USB 3.0 devices are detected successfully.

Change-Id: I5e4ab80c8d613a9cdeaba0501024f1c13c9549de
Signed-off-by: Ross Zwisler <zwisler@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1872997
Reviewed-by: Jon Flatley <jflat@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>