FROMGIT: rtw88: add power tracking support
The temperature of the chip can affect the output power
of the RF components. Hence driver requires to compensate
the power by adjusting the power index recorded in the
power swing table.

And if the difference of current thermal value to the
default thermal value exceeds a threshold, the RF IQK
should be triggered to re-calibrate the characteristics
of the RF components, to keep the output IQ vectors of
the RF components orthogonal enough.

Signed-off-by: Tzu-En Huang <tehuang@realtek.com>
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Reviewed-by: Chris Chiu <chiu@endlessm.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
(cherry picked from commit c97ee3e0bea29827f4b44276fc792bd32977edb0
 git://git.kernel.org/pub/scm/linux/kernel/git/kvalo/wireless-drivers-next.git master)

BUG=b:124400293
TEST=raise the temperature of the device, and see if power is expected

Change-Id: I4416087df33c592f750bf47b862e8d8c78de05c9
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1710429
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>