FROMGIT: rtw88: Enable 802.11ac beamformee support
Enable MU-MIMO transmit beamformee support for chipset 8822b and 8822c.

If the driver is in station mode and associated with an AP, and the
capabilities of both meet the requirement of beamforming, driver will
run as a beamformee and the corresponding chip settings will be set.

In addition, module parameter support_bf is added to enable or disable
beamforming. Sometimes driver will need to disable for inter-operate
issues, and it would be easier for driver to debug.

Signed-off-by: Tzu-En Huang <tehuang@realtek.com>
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
(cherry picked from commit 0bd9557341b7fb44bf591921d7feb4dcf4f4bb52
 git://git.kernel.org/pub/scm/linux/kernel/git/kvalo/wireless-drivers-next.git master)

BUG=b:124400293
TEST=Connect to an AP with beamforming, and see if there is sounding
     protocol exchange

Change-Id: I935c9ae076624c6ee0da1ac94d4c8e40522ce1cf
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1710435
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>