FROMGIT: ASoC: rt5677: Enable jack detect while DSP is running
Before a hotword is detected, GPIO1 pin is configured as IRQ
output so that jack detect works. When a hotword is detected,
the DSP firmware configures the GPIO1 pin as GPIO1 and
drives a 1. rt5677_irq() is called after a rising edge on
the GPIO1 pin, due to either jack detect event or hotword
event, or both. All possible events are checked and handled
in rt5677_irq() where GPIO1 pin is configured back to IRQ
output if a hotword is detected.

Signed-off-by: Ben Zhang <benzh@chromium.org>
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Link: https://lore.kernel.org/r/20191106011335.223061-4-cujomalainey@chromium.org
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 21c00e5df4397870ee835c974bf50570f9d24253
 https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
 for-5.5)

TEST=test jack detect and hotwording after series is applied
BUG=b:139552675

Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Change-Id: I8421e54887ce5a95adc2eb50c62fb5476952fde0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1918536