FROMGIT: ASoC: rt5677: Allow VAD to be shut on/off at all times
Due to limitations of the clocking configuration, we have no way of
scheduling our hibernation before the bdw dsp hibernates. This causes
issues when the system suspends with an open stream. We need userspace
to toggle the kcontrol before we are suspended so that any writes on
suspend are not lost and we don't corrupt the regmap.

Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Link: https://lore.kernel.org/r/20191106011335.223061-9-cujomalainey@chromium.org
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 395f02ef210e6f7dad2ca0efc16ffaea4a9ed030
 https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
 for-5.5)

TEST=test jack detect and hotwording after series is applied
BUG=b:139552675

Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Change-Id: If7375ae796ffdef4553b325905002bc1514120f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1918821