CHROMIUM: mm: add direct aging to kstaled
We have removed direct aging from kstaled when we removed pmd list
walk. Now it's time to add it back. Besides supporting vma, we now
also call page_referenced_one() for thp and shared pages.

If we try to reclaim a thp without calling page_referenced_one(), we
will fail to unmap it. But we also split it in the process, and this
side effect might be worse than paying for the rmap overhead.

For a shared page, it's similar. If one of the mappings is accessed,
we will fail to unmap it after we have unmapped the rest. So it
might not be worth unmapping any of its mappings at all.

I didn't observe any of the problems, but my test shows having this
patch doesn't have measurable rmap overhead either.

BUG=b:123039911
TEST=Ran platform.MemoryPressure and platform.MemoryStress

Change-Id: I9d1091f4992d22e21c332ec73c88bc0e94035ab9
Signed-off-by: Yu Zhao <yuzhao@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1911379
Reviewed-by: Salman Qazi <sqazi@google.com>
Commit-Queue: Yu Zhao <yuzhao@chromium.org>
Tested-by: Yu Zhao <yuzhao@chromium.org>